+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -1,53 +1,16 @@
 package roomescape.repository;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 import roomescape.controller.dto.ReservationTimeCreateRequest;
 import roomescape.domain.ReservationTime;
-import roomescape.util.CustomDateTimeFormatter;
 
-@Repository
-public class ReservationTimeRepository {
+public interface ReservationTimeRepository {
+    List<ReservationTime> findReservationTimes();
 
-    private final JdbcTemplate jdbcTemplate;
+    Long createReservationTime(ReservationTimeCreateRequest reservationTimeCreateRequest);
 
-    public ReservationTimeRepository(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
+    Optional<ReservationTime> findReservationTimeById(Long createdReservationTimeId);
 
-    public List<ReservationTime> findReservationTimes() {
-        return jdbcTemplate.query(
-                "SELECT * FROM reservation_time"
-                , (rs, rowNum) -> new ReservationTime(
-                        rs.getLong("id"),
-                        CustomDateTimeFormatter.getLocalTime(rs.getString("start_at"))
-                )
-        );
-    }
-
-    public Long createReservationTime(ReservationTimeCreateRequest reservationTimeCreateRequest) {
-        SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate).withTableName("reservation_time")
-                .usingGeneratedKeyColumns("id");
-        return simpleJdbcInsert.executeAndReturnKey(Map.of(
-                "start_at", reservationTimeCreateRequest.startAt()
-        )).longValue();
-    }
-
-    public Optional<ReservationTime> findReservationTimeById(Long createdReservationTimeId) {
-        return Optional.ofNullable(jdbcTemplate.queryForObject(
-                "SELECT * FROM reservation_time WHERE id = ?",
-                (rs, rowNum) -> new ReservationTime(
-                        rs.getLong("id"),
-                        CustomDateTimeFormatter.getLocalTime(rs.getString("start_at")
-                        )),
-                createdReservationTimeId));
-    }
-
-    public void deleteReservationTimeById(Long id) {
-        jdbcTemplate.update("DELETE FROM reservation_time WHERE id = ?", id);
-    }
+    void deleteReservationTimeById(Long id);
 }
